Do not attempt to create two subscriptions at once
The "subscribe if not subscribed" logic needs to work around Stripe's
tendency to get into a confused state if a subscription changes too
quickly (officially, within a 5-minute window; in practice, within a few
seconds).

Since we look up the phase and then set a subscription based on that
result, it was running into a state where the check happened twice,
resulting in two subscriptions, triggering the bad state about 50% of
the time.

The hacky workaround for now is this simple inflight detector pattern.
Put the promise in a Map, and if we see it's there, just return the same
one.